don't save config state for category load
