+++ libs/category/src/effects/category.effects.spec.ts
@@ -14,22 +14,12 @@ import {
 	DaffCategoryLoad,
 	DaffCategoryLoadSuccess,
 	DaffCategoryLoadFailure,
-	DaffChangeCategoryPageSize,
-	DaffChangeCategoryCurrentPage,
-	DaffChangeCategoryFilters,
-	DaffChangeCategorySortingOption,
-	DaffToggleCategoryFilter,
-  DaffCategoryPageLoad,
-  DaffCategoryPageLoadFailure,
-  DaffCategoryPageLoadSuccess
 } from '../actions/category.actions';
 import { DaffCategory } from '../models/category';
 import { DaffCategoryServiceInterface } from '../drivers/interfaces/category-service.interface';
 import { DaffCategoryDriver } from '../drivers/injection-tokens/category-driver.token';
 import { DaffCategoryPageConfigurationState } from '../models/category-page-configuration-state';
-import { DaffSortDirectionEnum, DaffCategoryRequest } from '../models/requests/category-request';
-import { DaffCategoryFilterEqualRequest } from '../models/requests/filter-request';
-import { DaffCategoryFilterType } from '../models/category-filter-base';
+import { DaffCategoryRequest } from '../models/requests/category-request';
 import { daffCategoryReducers } from '../reducers/category-reducers';
 
 class MockCategoryDriver implements DaffCategoryServiceInterface<DaffCategoryRequest, DaffCategory, DaffCategoryPageConfigurationState<DaffCategoryRequest>, DaffProduct> {
@@ -106,88 +96,14 @@ describe('DaffCategoryEffects', () => {
     expect(effects).toBeTruthy();
   });
 
-  describe('processCategoryGetRequest', () => {
-
-		let expected;
-		let categoryLoadAction;
-
-		beforeEach(() => {
- 			categoryLoadAction = new DaffCategoryLoad({id: stubCategory.id});
-		})
-
-    describe('when the call to CategoryService is successful', () => {
-
-      beforeEach(() => {
-        actions$ = hot('--a', { a: categoryLoadAction });
-      });
-
-      it('should dispatch a DaffCategoryLoadSuccess and a DaffProductGridLoadSuccess action', () => {
-				spyOn(daffCategoryDriver, 'get').and.returnValue(of({
-					category: stubCategory,
-					categoryPageConfigurationState: stubCategoryPageConfigurationState,
-					products: stubProducts
-				}));
-
-        expected = cold('--(ab)', { a: productGridLoadSuccessAction, b: categoryLoadSuccessAction });
-        expect(effects.loadCategory$).toBeObservable(expected);
-      });
-    });
-
-    describe('when the call to CategoryService fails', () => {
-
-      beforeEach(() => {
-        const error = 'Failed to load the category';
-				const response = cold('#', {}, error);
-				spyOn(daffCategoryDriver, 'get').and.returnValue(response);
-        const categoryLoadFailureAction = new DaffCategoryLoadFailure(error);
-        actions$ = hot('--a', { a: categoryLoadAction });
-        expected = cold('--b', { b: categoryLoadFailureAction });
-      });
-
-      it('should dispatch a CategoryLoadFailure action', () => {
-        expect(effects.loadCategory$).toBeObservable(expected);
-      });
-    });
-	});
-
-	describe('validateFilters', () => {
-
-		it('should throw an error when an invalid range type format is given', () => {
-			expect(() => {
-				effects.validateFilters([{
-					name: 'price',
-					value: ['invalidFormat'],
-					type: DaffCategoryFilterType.Range
-				}])
-			}).toThrowError('Category range filter is in an invalid format. Should be **-**');
-		});
-
-		it('should throw an error when two filters share the same name', () => {
-			expect(() => {
-				effects.validateFilters([{
-					name: 'name',
-					value: ['value'],
-					type: DaffCategoryFilterType.Equal
-				},
-				{
-					name: 'name',
-					value: ['value2'],
-					type: DaffCategoryFilterType.Equal
-				}
-			])
-			}).toThrowError('More than one category filter of the same name exists. These should' +
-						' be combined into a single filter action with multiple values.');
-		});
-	});
-
-  describe('when CategoryPageLoadAction is triggered', () => {
-    let categoryPageLoadSuccessAction: DaffCategoryPageLoadSuccess;
+  describe('when CategoryLoadAction is triggered', () => {
+    let categoryPageLoadSuccessAction: DaffCategoryLoadSuccess;
 		let expected;
 		let categoryPageLoadAction;
 
 		beforeEach(() => {
-       categoryPageLoadAction = new DaffCategoryPageLoad({id: stubCategory.id});
-       categoryPageLoadSuccessAction = new DaffCategoryPageLoadSuccess({
+       categoryPageLoadAction = new DaffCategoryLoad({id: stubCategory.id});
+       categoryPageLoadSuccessAction = new DaffCategoryLoadSuccess({
         category: stubCategory,
         categoryPageConfigurationState: stubCategoryPageConfigurationState,
         products: stubProducts
@@ -197,7 +113,7 @@ describe('DaffCategoryEffects', () => {
 		});
 
     describe('when the call to CategoryService is successful', () => {
-      it('should dispatch a DaffCategoryPageLoadSuccess and a DaffProductGridLoadSuccess action', () => {
+      it('should dispatch a DaffCategoryLoadSuccess and a DaffProductGridLoadSuccess action', () => {
 				spyOn(daffCategoryDriver, 'get').and.returnValue(of({
 					category: stubCategory,
 					categoryPageConfigurationState: stubCategoryPageConfigurationState,
@@ -205,7 +121,7 @@ describe('DaffCategoryEffects', () => {
 				}));
 
         expected = cold('--(ab)', { a: productGridLoadSuccessAction, b: categoryPageLoadSuccessAction });
-        expect(effects.loadCategoryPage$).toBeObservable(expected);
+        expect(effects.loadCategory$).toBeObservable(expected);
       });
     });
 
@@ -215,12 +131,12 @@ describe('DaffCategoryEffects', () => {
         const error = 'Failed to load the category';
 				const response = cold('#', {}, error);
 				spyOn(daffCategoryDriver, 'get').and.returnValue(response);
-        const categoryPageLoadFailureAction = new DaffCategoryPageLoadFailure(error);
+        const categoryPageLoadFailureAction = new DaffCategoryLoadFailure(error);
         expected = cold('--b', { b: categoryPageLoadFailureAction });
       });
 
-      it('should dispatch a CategoryPageLoadFailure action', () => {
-        expect(effects.loadCategoryPage$).toBeObservable(expected);
+      it('should dispatch a CategoryLoadFailure action', () => {
+        expect(effects.loadCategory$).toBeObservable(expected);
       });
     });
 
@@ -232,161 +148,9 @@ describe('DaffCategoryEffects', () => {
 			}));
 
       expected = cold('--(ab)', { a: productGridLoadSuccessAction, b: categoryPageLoadSuccessAction });
-      expect(effects.loadCategoryPage$).toBeObservable(expected);
-
-      expect(daffCategoryDriver.get).toHaveBeenCalledWith({ id: stubCategory.id });
-    });
-  });
-
-  describe('when CategoryLoadAction is triggered', () => {
-
-    let expected;
-
-    it('should call get category with only an id', () => {
-			spyOn(daffCategoryDriver, 'get').and.returnValue(of({
-				category: stubCategory,
-				categoryPageConfigurationState: stubCategoryPageConfigurationState,
-				products: stubProducts
-			}));
-      const categoryLoadAction = new DaffCategoryLoad({ id: stubCategory.id });
-      actions$ = hot('--a', { a: categoryLoadAction });
-
-      expected = cold('--(ab)', { a: productGridLoadSuccessAction, b: categoryLoadSuccessAction });
       expect(effects.loadCategory$).toBeObservable(expected);
 
       expect(daffCategoryDriver.get).toHaveBeenCalledWith({ id: stubCategory.id });
-    });
-	});
-
-  describe('when ChangeCategoryPageSize action is triggered', () => {
-
-    let expected;
-
-    it('should call get category with an id, page size, applied filters, and an applied sort option', () => {
-			spyOn(daffCategoryDriver, 'get').and.returnValue(of({
-				category: stubCategory,
-				categoryPageConfigurationState: stubCategoryPageConfigurationState,
-				products: stubProducts
-			}));
-      const changeCategoryPageSizeAction = new DaffChangeCategoryPageSize(3);
-      actions$ = hot('--a', { a: changeCategoryPageSizeAction });
-
-      expected = cold('--(ab)', { a: productGridLoadSuccessAction, b: categoryLoadSuccessAction });
-			expect(effects.changeCategoryPageSize$).toBeObservable(expected);
-      expect(daffCategoryDriver.get).toHaveBeenCalledWith({
-				...stubCategoryPageConfigurationState,
-				page_size: 3
-      });
-    });
-  });
-
-  describe('when ChangeCategoryCurrentPage action is triggered', () => {
-
-    let expected;
-
-    it('should call get category with every available parameter', () => {
-			spyOn(daffCategoryDriver, 'get').and.returnValue(of({
-				category: stubCategory,
-				categoryPageConfigurationState: stubCategoryPageConfigurationState,
-				products: stubProducts
-			}));
-      const changeCategoryCurrentPageAction = new DaffChangeCategoryCurrentPage(3);
-      actions$ = hot('--a', { a: changeCategoryCurrentPageAction });
-
-			expected = cold('--(ab)', { a: productGridLoadSuccessAction, b: categoryLoadSuccessAction });
-			expect(effects.changeCategoryCurrentPage$).toBeObservable(expected);
-			expect(daffCategoryDriver.get).toHaveBeenCalledWith({
-				...stubCategoryPageConfigurationState,
-				current_page: 3
-      });
-    });
-  });
-
-  describe('when ChangeCategoryFiltersAction is triggered', () => {
-
-    let expected;
-
-    it('should call get category with an id, page size, applied filters, and an applied sorting option', () => {
-			spyOn(daffCategoryDriver, 'get').and.returnValue(of({
-				category: stubCategory,
-				categoryPageConfigurationState: stubCategoryPageConfigurationState,
-				products: stubProducts
-			}));
-      const changeCategoryFiltersAction = new DaffChangeCategoryFilters([{
-				name: 'name',
-				type: DaffCategoryFilterType.Equal,
-				value: ['value']
-			}]);
-      actions$ = hot('--a', { a: changeCategoryFiltersAction });
-
-			expected = cold('--(ab)', { a: productGridLoadSuccessAction, b: categoryLoadSuccessAction });
-			expect(effects.changeCategoryFilters$).toBeObservable(expected);
-			expect(daffCategoryDriver.get).toHaveBeenCalledWith({
-				...stubCategoryPageConfigurationState,
-				filter_requests: [{
-					name: 'name',
-					type: DaffCategoryFilterType.Equal,
-					value: ['value']
-				}]
-      });
-    });
-	});
-
-  describe('when ToggleCategoryFilterAction is triggered', () => {
-
-    let expected;
-
-    it('should call get category with an id, page size, applied filters, and an applied sorting option', () => {
-			spyOn(daffCategoryDriver, 'get').and.returnValue(of({
-				category: stubCategory,
-				categoryPageConfigurationState: stubCategoryPageConfigurationState,
-				products: stubProducts
-			}));
-			const appliedFilter: DaffCategoryFilterEqualRequest = {
-				name: 'name',
-				type: DaffCategoryFilterType.Equal,
-				value: ['value']
-			}
-      const toggleCategoryFilterAction = new DaffToggleCategoryFilter({
-				name: 'name',
-				type: DaffCategoryFilterType.Equal,
-				value: 'value'
-			});
-			store.dispatch(toggleCategoryFilterAction);
-      actions$ = hot('--a', { a: toggleCategoryFilterAction });
-
-			expected = cold('--(ab)', { a: productGridLoadSuccessAction, b: categoryLoadSuccessAction });
-			expect(effects.toggleCategoryFilter$).toBeObservable(expected);
-			expect(daffCategoryDriver.get).toHaveBeenCalledWith({
-        ...stubCategoryPageConfigurationState,
-				filter_requests: [appliedFilter]
-      });
-    });
-	});
-
-  describe('when ChangeCategorySortAction is triggered', () => {
-
-    let expected;
-
-    it('should call get category with an id, page size, applied filters, and an applied sorting option', () => {
-			spyOn(daffCategoryDriver, 'get').and.returnValue(of({
-				category: stubCategory,
-				categoryPageConfigurationState: stubCategoryPageConfigurationState,
-				products: stubProducts
-			}));
-      const changeCategorySortingOption = new DaffChangeCategorySortingOption({
-				option: 'option',
-				direction: DaffSortDirectionEnum.Ascending
-			});
-      actions$ = hot('--a', { a: changeCategorySortingOption });
-
-			expected = cold('--(ab)', { a: productGridLoadSuccessAction, b: categoryLoadSuccessAction });
-			expect(effects.changeCategorySort$).toBeObservable(expected);
-			expect(daffCategoryDriver.get).toHaveBeenCalledWith({
-				...stubCategoryPageConfigurationState,
-				applied_sort_direction: DaffSortDirectionEnum.Ascending,
-				applied_sort_option: 'option'
-      });
     });
   });
 });
