+++ libs/category/src/reducers/category/category.reducer.ts
@@ -29,11 +29,17 @@ const initialState: DaffCategoryReducerState<any, any> = {
 export function daffCategoryReducer<T extends DaffCategoryRequest, U extends DaffGenericCategory<U>, V extends DaffCategoryPageConfigurationState<T>, W extends DaffProduct>
 	(state = initialState, action: DaffCategoryActions<T, U, V, W>): DaffCategoryReducerState<T, V> {
   switch (action.type) {
+    case DaffCategoryActionTypes.CategoryLoadAction:
+      return {
+        ...state,
+				categoryLoading: true,
+				productsLoading: true,
+      };
+
     // This reducer must assume the call will be successful, and immediately set the applied filters to state, because the
     // GetCategory magento call doesn't return currently applied filters. If there is a bug where the wrong filters are somehow
     // applied by Magento, then this will result in a bug. Until Magento returns applied filters with a category call, this is
     // unavoidable.
-    case DaffCategoryActionTypes.CategoryLoadAction:
     case DaffCategoryActionTypes.CategoryPageLoadAction:
       return {
         ...state,
@@ -92,7 +98,6 @@ export function daffCategoryReducer<T extends DaffCategoryRequest, U extends Daf
 			}
     // This reducer cannot spread over state, because this would wipe out the applied filters on state. Applied filters are not
     // set here for reasons stated above.
-    case DaffCategoryActionTypes.CategoryLoadSuccessAction:
     case DaffCategoryActionTypes.CategoryPageLoadSuccessAction:
       return {
         ...state,
